+++ vscode-extension/src/objectwizards/syntaxbuilders/alEnumExtSyntaxBuilder.ts
@@ -1,6 +1,7 @@
 import * as vscode from 'vscode';
 import { ALEnumExtWizardData } from "../wizards/alEnumExtWizardData";
 import { ALSyntaxWriter } from "../../allanguage/alSyntaxWriter";
+import { ALSyntaxHelper } from '../../allanguage/alSyntaxHelper';
 
 export class ALEnumExtSyntaxBuilder {
     
@@ -15,20 +16,17 @@ export class ALEnumExtSyntaxBuilder {
         writer.writeLine("");
 
         if ((data.valueList) && (data.valueList != "")) {
-            let values : string[] = data.valueList.split(",");
-            let captions : string[];
-            if ((data.captionList) && (data.captionList != ""))
-                captions = data.captionList.split(",");
-            else
-                captions = [];
+            let values : string[] = ALSyntaxHelper.splitNamesList(data.valueList);
+            let captions : string[] = ALSyntaxHelper.splitNamesList(data.captionList);
             let valueId = data.firstValueId;
             if ((values) && (values.length > 0)) {
                 for (let i = 0; i < values.length; i++) {
                     writer.writeStartNameSourceBlock("value", valueId.toString(), writer.encodeName(values[i]));
-                    if (captions.length > i)
+                    if (captions.length > i) {
                         writer.writeProperty("Caption", writer.encodeString(captions[i]));
-                    else
+                    } else {
                         writer.writeProperty("Caption", writer.encodeString(values[i]));
+                    }
                     writer.writeEndBlock();
                     valueId++;
                 }